+++ Meta/gn/secondary/Userland/Libraries/LibGfx/BUILD.gn
@@ -97,6 +97,7 @@ shared_library("LibGfx") {
     "ImageFormats/WebPLoader.cpp",
     "ImageFormats/WebPLoaderLossless.cpp",
     "ImageFormats/WebPLoaderLossy.cpp",
+    "ImageFormats/WebPSharedLossless.cpp",
     "ImageFormats/WebPWriter.cpp",
     "ImageFormats/WebPWriterLossless.cpp",
     "ImmutableBitmap.cpp",
